+++ vector/src/main/java/im/vector/app/features/call/VectorCallActivity.kt
@@ -23,7 +23,6 @@ import android.content.Intent.FLAG_ACTIVITY_CLEAR_TOP
 import android.os.Build
 import android.os.Bundle
 import android.os.Parcelable
-import android.view.KeyEvent
 import android.view.View
 import android.view.Window
 import android.view.WindowManager
@@ -332,20 +331,6 @@ class VectorCallActivity : VectorBaseActivity(), CallControlsView.InteractionLis
             callViewModel.handle(VectorCallViewActions.ToggleCamera)
         }
         surfaceRenderersAreInitialized = true
-    }
-
-    override fun onKeyDown(keyCode: Int, event: KeyEvent?): Boolean {
-        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.LOLLIPOP) {
-            // for newer version, it will be passed automatically to active media session
-            // in call service
-            when (keyCode) {
-                KeyEvent.KEYCODE_HEADSETHOOK -> {
-                    callViewModel.handle(VectorCallViewActions.HeadSetButtonPressed)
-                    return true
-                }
-            }
-        }
-        return super.onKeyDown(keyCode, event)
     }
 
     private fun handleViewEvents(event: VectorCallViewEvents?) {