+++ vector/src/main/java/im/vector/app/features/webview/VectorWebViewClient.kt
@@ -26,7 +26,6 @@ import android.webkit.WebResourceRequest
 import android.webkit.WebResourceResponse
 import android.webkit.WebView
 import android.webkit.WebViewClient
-import androidx.annotation.RequiresApi
 
 /**
  * This class inherits from WebViewClient. It has to be used with a WebView.
@@ -58,7 +57,6 @@ class VectorWebViewClient(private val eventListener: WebViewEventListener) : Web
         }
     }
 
-    @RequiresApi(Build.VERSION_CODES.LOLLIPOP)
     override fun onReceivedHttpError(view: WebView, request: WebResourceRequest, errorResponse: WebResourceResponse) {
         super.onReceivedHttpError(view, request, errorResponse)
         eventListener.onHttpError(request.url.toString(),