+++ CHANGES.md
@@ -33,6 +33,7 @@ Other changes:
  - Rename package `im.vector.riotx.attachmentviewer` to `im.vector.lib.attachmentviewer`
  - Rename package `im.vector.riotx.multipicker` to `im.vector.lib.multipicker`
  - Rename package `im.vector.riotx` to `im.vector.app`
+ - Remove old code that was used on devices with api level <21
  - Add Official Gradle Wrapper Validation Action
 
 Changes in Element 1.0.4 (2020-08-03)